+++ CHANGELOG.md
@@ -27,6 +27,8 @@ MyCMS::InitDatabase class to read database configuration from (dist/)phinx.yml
 - phpcbf automatically fixed style in styles/admin.css, dist/styles/admin.css, dist/scripts/admin-specific.js, dist/scripts/admin.js, dist/classes/Admin.php
 - assert array as argument of array_shift in LogMysqli::fetchAndReindex
 - dist/images/favicon/manifest.json 12x fixed error  Unnecessary escape character: \/  jsonc/no-useless-escape
+- fix classes/MyAdmin.php line 640: Ternary operator condition is always true.
+- fix classes/ProjectCommon.php line 150: Parameter #1 $pattern of function WorkOfStan\MyCMS\ThrowableFunctions\preg_replaceString expects array<string>|string, array<int, (int|string)> given.
 
 ### `Security` in case of vulnerabilities
 - dist/.htaccess: no markdown file can be available on web